+++ core/src/main/java/org/mule/runtime/core/internal/rx/FluxSinkRecorder.java
@@ -68,20 +68,4 @@ public class FluxSinkRecorder<T> implements Consumer<FluxSink<T>> {
       fluxSink.error(error);
     }
   }
-
-  public void complete() {
-    boolean present = true;
-    synchronized (this) {
-      if (fluxSink == null) {
-        present = false;
-        bufferedEvents.add(() -> {
-          fluxSink.complete();
-        });
-      }
-    }
-
-    if (present) {
-      fluxSink.complete();
-    }
-  }
 }
